requirements.txt
django
djangorestframework
pillow
requests
djangorestframework-simplejwt
mysqlclient